api/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3
pygments===2.3.1
django-cors-headers===3.10.0
